# app/src/docs/v1.api-spec.yaml
...
openapi: 3.0.2
info:
  version: 1.0.0
  title: GetOK API
  description: >-
    This API automates the process for getting setup to use common services
    published on https://apistore.nrs.gov.bc.ca/store/
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
  contact:
    name: NR Common Service Showcase
    email: NR.CommonServiceShowcase@gov.bc.ca
servers:
  - url: /api/v1
    description: This Server
security:
  - bearerAuth: []
  - OpenID: []
paths:
...
  "/users/{keycloakId}/acronyms":
    get:
      summary: Returns acronyms associated with a user
      operationId: userAcronymGet
      tags:
        - Users
      parameters:
        - name: keycloakId
          in: path
          description: Keycloak user UUID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A list of associated acronyms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAcronymList"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  "/users/{keycloakId}/acronyms/clients":
    get:
      summary: Returns service clients for all realms for all acronyms associated with a user
      operationId: userAcronymClientGet
      tags:
        - Users
      parameters:
        - name: keycloakId
          in: path
          description: Keycloak user UUID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A list of associated acronyms' clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcronymClientsList"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OpenID:
      type: openIdConnect
      openIdConnectUrl: "https://example.com/.well-known/openid-configuration"
  schemas:
...
    AcronymClients:
      type: object
      properties:
        acronym:
          type: string
          example: MSSC
        dev:
          $ref: "#/components/schemas/KcClientDetail"
        test:
          $ref: "#/components/schemas/KcClientDetail"
        prod:
          $ref: "#/components/schemas/KcClientDetail"
    AcronymClientsList:
      type: array
      items:
        $ref: "#/components/schemas/AcronymClients"
...
    Error:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            status:
              example: 500
            title:
              example: Internal Server Error
            type:
              example: "https://httpstatuses.com/500"
    KcClientDetail:
      type: object
      required:
        - environment
        - id
        - clientId
        - enabled
        - name
        - description
        - commonServiceRoles
      properties:
        environment:
          type: string
          enum:
            - DEV
            - TEST
            - PROD
          description: >-
            Which KeyCloak environment the client is in - one of DEV, TEST or
            PROD
          example: TEST
        id:
          type: string
          example: d288f328-3d0e-4bd0-8d89-8beccdace260
        clientId:
          type: string
          example: ZZZZ_SERVICE_CLIENT
        enabled:
          type: boolean
          example: true
        name:
          type: string
          example: ZZZ Application
        description:
          type: string
          example: This application does some work
        serviceAccountEmail:
          type: string
          example: bob.alice@gov.bc.ca
        commonServiceRoles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
          description: An array of Common Service roles the client currently has
...
    NotFound:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            status:
              example: 404
            title:
              example: Not Found
            type:
              example: "https://httpstatuses.com/404"
    Problem:
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          description: "What type of problem, link to explanation of problem"
        title:
          type: string
          description: "Title of problem, generally the Http Status Code description"
        status:
          type: string
          description: The Http Status code
        detail:
          type: string
          description: Short description of why this problem was raised.
    Role:
      type: object
      required:
        - roleName
        - description
      properties:
        roleName:
          type: string
          example: EMAILER
        description:
          type: string
          example: This role allows email sending via CHES
    UserAcronymList:
      type: array
      items:
        $ref: "#/components/schemas/UserAcronymDetail"
      description: An array of acronym detail items
    UserAcronymDetail:
      type: object
      properties:
        acronym:
          type: string
          description: The application acronym
          example: MSSC
        owner:
          type: boolean
          description: User is an owner of this acronym
          example: false
        createdAt:
          type: string
          description: The datestamp the user->acronym mapping was created
          example: "2019-08-22T18:43:07.393Z"
...
    ValidationError:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          required:
            - errors
          properties:
            errors:
              type: array
              items:
                type: object
                required:
                  - message
                properties:
                  value:
                    type: object
                    description: Contents of the field that was in error.
                    example: utf-8x
                  message:
                    type: string
                    description: The error message for the field.
                    example: Invalid value `encoding`.
            status:
              example: 422
            title:
              example: Unprocessable Entity
            type:
              example: "https://httpstatuses.com/422"
...
  responses:
...
    Forbidden:
      description: Lack required role to perform action
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFound"
    UnauthorizedError:
      description: Access token is missing or invalid
    UnprocessableEntity:
      description: >-
        The server was unable to process the contained instructions. Generally
        validation error(s).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"